formatter - refactor
